запущено под x64

+++ runtime/include/pgc/result.hpp
@@ -23,8 +23,8 @@ namespace pgc
 		friend class Statement;
 		Result(ResultImplPtr impl);
 
-		bool fetchHelper(size_t rowIdx, size_t colIdx, int typIdx, void *data);
-		bool fetchHelper(size_t rowIdx, const char *colName, int typIdx, void *data);
+		bool fetchHelper(int rowIdx, int colIdx, int typIdx, void *data);
+		bool fetchHelper(int rowIdx, const char *colName, int typIdx, void *data);
 
 	public:
 		~Result();
@@ -32,24 +32,24 @@ namespace pgc
 		EExecStatus status();
 		Result &throwIfError();
 
-		size_t rows();
-		template <class T> Result &fetch(size_t rowIdx, size_t colIdx, T &v);
-		template <class T> Result &fetch(size_t rowIdx, const char *colName, T &v);
+		int rows();
+		template <class T> Result &fetch(int rowIdx, int colIdx, T &v);
+		template <class T> Result &fetch(int rowIdx, const char *colName, T &v);
 
-		bool isNull(size_t rowIdx, size_t colIdx);
-		bool isNull(size_t rowIdx, const char *colName);
+		bool isNull(int rowIdx, int colIdx);
+		bool isNull(int rowIdx, const char *colName);
 	};
 	typedef boost::shared_ptr<Result> ResultPtr;
 
 	//////////////////////////////////////////////////////////////////////////
-	template <class T> Result &Result::fetch(size_t rowIdx, size_t colIdx, T &v)
+	template <class T> Result &Result::fetch(int rowIdx, int colIdx, T &v)
 	{
 		fetchHelper(rowIdx, colIdx, CppDataType<T>::cdt_index, &v);
 		return *this;
 	}
 
 	//////////////////////////////////////////////////////////////////////////
-	template <class T> Result &Result::fetch(size_t rowIdx, const char *colName, T &v)
+	template <class T> Result &Result::fetch(int rowIdx, const char *colName, T &v)
 	{
 		fetchHelper(rowIdx, colName, CppDataType<T>::cdt_index, &v);
 		return *this;
